Getting closer to running quickChick
This version of the file does not build yet, but it is close. I suspect
one of the main problems with the parser, in addition to some complexity
in the file, is that the QuickChick comment notation for running tests
in the tool does not like QuickChick commands in the way they are
written now.